create errorBoundary placeholder for future UI designs when app errors
